server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
cffi==1.15.1
channels==3.0.4
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.1
daphne==3.0.2
Django==3.2.8
django-cors-headers==3.8.0
django-shortuuidfield==0.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
multidict==6.0.4
openai==0.27.4
Pillow==8.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
pytz==2023.3
requests==2.28.2
service-identity==21.1.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.3
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
zope.interface==6.0
